refactored deletepolicy functoin

+++ src/app/policy-list/policy-list.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { PolicyService } from '../policy.service';
 import { policydataitem } from '../model/policydataitem';
+import { catchError } from 'rxjs';
 
 @Component({
   selector: 'app-policy-list',
@@ -8,69 +9,65 @@ import { policydataitem } from '../model/policydataitem';
   styleUrls: ['./policy-list.component.css']
 })
 
-export class PolicyListComponent 
-{
+export class PolicyListComponent {
   policyArray!: policydataitem[];
 
   policies: any[] = [];
   constructor(private policyService: PolicyService) { }
 
-    ngOnInit()
-      {
-      this.getPolicyData()
-    }
+  ngOnInit() {
+    this.getPolicyData()
+  }
 
-    getPolicyData (): void
-    {
+  getPolicyData(): void {
 
-      this.policyService.getPolicies().subscribe
+    this.policyService.getPolicies().subscribe
       (
-          (data: any): void =>
-          {
+        (data: any): void => {
           console.log(data);
           this.policies = data;
           this.policyArray = data;
-          }
+        }
       )
 
+  }
+
+  viewButton(policy: policydataitem): void {
+    console.log(policy.id);
+  }
+
+  // navigateToPolicyDetail( policy : policydataitem ) :void
+  // {
+
+  //   const policyIdString = policy?.id.toString();
+
+  //   this.router.navigate(['/policy-detail', policyIdString]);
+
+  // }
+
+  deleteButton(policy: policydataitem): void {
+
+    // filter policy list
+    const policyFiltered = this.filterPoliyList(policy);
+
+    // when filtered
+    try {
+          if (!policyFiltered) this.policyService.deletePolicy1(policy).subscribe();
+    } catch (error) {
+      throw new Error("eror while deletion")
     }
-
-    viewButton(policy : policydataitem):void
-    {
-      console.log(policy.id);
-    }
-      
-    // navigateToPolicyDetail( policy : policydataitem ) :void
-    // {
-
-    //   const policyIdString = policy?.id.toString();
-
-    //   this.router.navigate(['/policy-detail', policyIdString]);
-
-    // }
     
-    deleteButton(policy : policydataitem): void 
-    {      
-      const policyFiltered = this.policyArray?.filter(p => p !== policy);
-      
-      if (!policyFiltered)
-      {    
-        
-        this.policyService.deletePolicy1(policy).subscribe();
-        console.log('inside - False ');
-      }
+    //Data bind
+    this.DataBind(policyFiltered);
 
-      if (policyFiltered)
-      {
-      
-        this.policyService.deletePolicy1(policy).subscribe();
-        console.log('inside  - True ');
-      }
-
-      this.policyArray = policyFiltered;
+  }
 
 
-    }
+  private DataBind(policyFiltered: policydataitem[]) {
+    this.policyArray = policyFiltered;
+  }
 
- 
+  private filterPoliyList(policy: policydataitem) {
+    return this.policyArray?.filter(p => p !== policy);
+  }
 }
